refactor: centralize solver base classes
