+++ src/dpf2/core/bases.py
@@ -1,0 +1,44 @@
+"""Abstract base classes for solver components.
+
+This file provides canonical interfaces for plasma, circuit and
+diagnostics modules.  Other copies in the repository should import
+from here to avoid duplication."""
+
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from typing import Any
+
+
+class PlasmaSolverBase(ABC):
+    """Interface for plasma solvers."""
+
+    @abstractmethod
+    def step(self, state: Any, dt: float) -> Any:
+        """Advance the plasma state by ``dt`` seconds."""
+        raise NotImplementedError
+
+
+class CircuitSolverBase(ABC):
+    """Interface for external circuit solvers."""
+
+    @abstractmethod
+    def step(self, current: float, voltage: float, dt: float) -> tuple[float, float]:
+        """Return updated (current, voltage) after ``dt`` seconds."""
+        raise NotImplementedError
+
+
+class DiagnosticsBase(ABC):
+    """Interface for simulation diagnostics."""
+
+    @abstractmethod
+    def record(self, state: Any, time: float) -> None:
+        """Record the simulation ``state`` at ``time``."""
+        raise NotImplementedError
+
+
+__all__ = [
+    "PlasmaSolverBase",
+    "CircuitSolverBase",
+    "DiagnosticsBase",
+]
